pyproject.toml
[tool.poetry]
name = "git-black"
version = "0.1.0"
description = ""
authors = ["Andy Teijelo <ateijelo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
gitpython = "^3.1.2"
click = "^7.1.2"
unidiff = "^0.6.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.2"
jinja2 = "^2.11.2"
black = "^19.10b0"
pytest = "^5.4.3"

[tool.poetry.scripts]
"git-black" = "git_black:cli"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
